install playwright to modal component
