+++ packages/modal/tools/webpack.config.js
